==> final_analysis/6.LMEM Model.ipynb <==
library(lme4)

# Vitals

data = read.csv('LMER/data_review_vitals_Physician_star.csv')

bmod <- lmer(ReviewScore~rcnt + past +
                         # rcnt*healthgrades + rcnt*ratemds + rcnt*vitals + rcnt*yelp +
                         # past*healthgrades + past*ratemds + past*vitals + past*yelp +
                         (rcnt + past-1 | NPI)-1,
              control = lmerControl(optimizer="bobyqa", optCtrl=list(maxfun=100000)),
              data=data)

summary(bmod)

data = read.csv('LMER/data_review_vitals_Physician_star_20Rew.csv')

bmod <- lmer(ReviewScore~rcnt + past +
                         # rcnt*healthgrades + rcnt*ratemds + rcnt*vitals + rcnt*yelp +
                         # past*healthgrades + past*ratemds + past*vitals + past*yelp +
                         (rcnt + past-1 | NPI)-1,
              control = lmerControl(optimizer="bobyqa", optCtrl=list(maxfun=100000)),
              data=data)

summary(bmod)

# HealthGrades

data = read.csv('LMER/data_Physician_star_review_healthgrades.csv')

bmod <- lmer(ReviewScore~rcnt + past +
                         # rcnt*healthgrades + rcnt*ratemds + rcnt*vitals + rcnt*yelp +
                         # past*healthgrades + past*ratemds + past*vitals + past*yelp +
                         (rcnt + past-1 | NPI)-1,
              control = lmerControl(optimizer="bobyqa", optCtrl=list(maxfun=100000)),
              data=data)

summary(bmod)

data = read.csv('LMER/data_Physician_star_20Rew_review_healthgrades.csv')

bmod <- lmer(ReviewScore~rcnt + past +
                         # rcnt*healthgrades + rcnt*ratemds + rcnt*vitals + rcnt*yelp +
                         # past*healthgrades + past*ratemds + past*vitals + past*yelp +
                         (rcnt + past-1 | NPI)-1,
              control = lmerControl(optimizer="bobyqa", optCtrl=list(maxfun=100000)),
              data=data)

summary(bmod)

# All with 50+ Reviews

data = read.csv('LMER/data_review_all_Physician_star_50Rew_new.csv')
bmod <- lmer(ReviewScore~rcnt + past +
                         # rcnt*healthgrades + rcnt*ratemds + rcnt*vitals + rcnt*yelp +
                         # past*healthgrades + past*ratemds + past*vitals + past*yelp +
                         (rcnt + past-1 | NPI)-1,
              control = lmerControl(optimizer="bobyqa", optCtrl=list(maxfun=100000)),
              data=data)

summary(bmod)


bmod <- lmer(ReviewScore~rcnt + past +
                         rcnt*healthgrades + rcnt*ratemds + rcnt*vitals + rcnt*yelp +
                         past*healthgrades + past*ratemds + past*vitals + past*yelp +
                         (rcnt + past-1 | NPI)-1,
              control = lmerControl(optimizer="bobyqa", optCtrl=list(maxfun=100000)),
              data=data)

summary(bmod)

# All with 20+ Reviews

data = read.csv('LMER/data_review_all_Physician_star_20Rew_new.csv')

bmod <- lmer(ReviewScore~rcnt + past +
                         # rcnt*healthgrades + rcnt*ratemds + rcnt*vitals + rcnt*yelp +
                         # past*healthgrades + past*ratemds + past*vitals + past*yelp +
                         (rcnt + past-1 | NPI)-1,
              control = lmerControl(optimizer="bobyqa", optCtrl=list(maxfun=100000)),
              data=data)

summary(bmod)


bmod <- lmer(ReviewScore~rcnt + past +
                         rcnt*healthgrades + rcnt*ratemds + rcnt*vitals + rcnt*yelp +
                         past*healthgrades + past*ratemds + past*vitals + past*yelp +
                         (rcnt + past-1 | NPI)-1,
              control = lmerControl(optimizer="bobyqa", optCtrl=list(maxfun=100000)),
              data=data)

summary(bmod)

# All with 10+ Reviews

data = read.csv('LMER/data_review_all_Physician_star_10Rew_new.csv')

bmod <- lmer(ReviewScore~rcnt + past +
                         # rcnt*healthgrades + rcnt*ratemds + rcnt*vitals + rcnt*yelp +
                         # past*healthgrades + past*ratemds + past*vitals + past*yelp +
                         (rcnt + past-1 | NPI)-1,
              control = lmerControl(optimizer="bobyqa", optCtrl=list(maxfun=100000)),
              data=data)

summary(bmod)

data = read.csv('LMER/data_review_all_Physician_star_10Rew_new.csv')

bmod <- lmer(ReviewScore~rcnt + past +
                         rcnt*healthgrades + rcnt*ratemds + rcnt*vitals + rcnt*yelp +
                         past*healthgrades + past*ratemds + past*vitals + past*yelp +
                         (rcnt + past-1 | NPI)-1,
              control = lmerControl(optimizer="bobyqa", optCtrl=list(maxfun=100000)),
              data=data)

summary(bmod)

# All

data = read.csv('LMER/data_review_all_Physician_star_new.csv')

bmod <- lmer(ReviewScore~rcnt + past +
                         # rcnt*healthgrades + rcnt*ratemds + rcnt*vitals + rcnt*yelp +
                         # past*healthgrades + past*ratemds + past*vitals + past*yelp +
                         (rcnt + past-1 | NPI)-1,
              control = lmerControl(optimizer="bobyqa", optCtrl=list(maxfun=100000)),
              data=data)

summary(bmod)

data = read.csv('LMER/data_review_all_Physician_star_new.csv')

bmod <- lmer(ReviewScore~rcnt + past +
                         rcnt*healthgrades + rcnt*ratemds + rcnt*vitals + rcnt*yelp +
                         past*healthgrades + past*ratemds + past*vitals + past*yelp +
                         (rcnt + past-1 | NPI)-1,
              control = lmerControl(optimizer="bobyqa", optCtrl=list(maxfun=100000)),
              data=data)

summary(bmod)
